% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeConstraints.R
\name{writeConstraints}
\alias{writeConstraints}
\title{Write constraint script}
\usage{
writeConstraints(targetdialect)
}
\arguments{
\item{targetdialect}{The dialect of the target database. Choices are "oracle", "postgresql", "pdw", "redshift", "impala", "netezza", "bigquery", "sql server"}
}
\description{
Write constraint script
}
