% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeDDL.R
\name{writeDDL}
\alias{writeDDL}
\title{Write DDL script}
\usage{
writeDDL(targetdialect)
}
\arguments{
\item{targetdialect}{The dialect of the target database. Choices are "oracle", "postgresql", "pdw", "redshift", "impala", "netezza", "bigquery", "sql server"}
}
\description{
Write DDL script
}
