% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeIndex.R
\name{writeIndex}
\alias{writeIndex}
\title{Write Index script}
\usage{
writeIndex(targetdialect)
}
\arguments{
\item{targetdialect}{The dialect of the target database. Choices are "oracle", "postgresql", "pdw", "redshift", "impala", "netezza", "bigquery", "sql server"}
}
\description{
Write Index script
}
