% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writePDF.R
\name{writePDF}
\alias{writePDF}
\title{Write PDF of CDM documentation}
\usage{
writePDF(mdFilesLocation, cdmVersion)
}
\arguments{
\item{mdFilesLocation}{The dialect of the target database. Choices are "oracle", "postgresql", "pdw", "redshift", "impala", "netezza", "bigquery", "sql server"}

\item{cdmVersion}{The version of the OMOP Common Data Model for which the pdf document is being written. It should be written without a decimal so v5.3 becomes v5_3}
}
\description{
Write PDF of CDM documentation
}
