% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writePrimaryKeys.R
\name{writePrimaryKeys}
\alias{writePrimaryKeys}
\title{Write primary key script}
\usage{
writePrimaryKeys(targetdialect)
}
\arguments{
\item{targetdialect}{The dialect of the target database. Choices are "oracle", "postgresql", "pdw", "redshift", "impala", "netezza", "bigquery", "sql server"}
}
\description{
Write primary key script
}
